+++ src/dipdup/cli.py
@@ -32,7 +32,18 @@ from dipdup.sys import set_up_process
 if TYPE_CHECKING:
     from dipdup.config import DipDupConfig
 
+ROOT_CONFIG = 'dipdup.yaml'
+CONFIG_RE = r'dipdup.*\.ya?ml'
 
+# NOTE: Do not try to load config for these commands as they don't need it
+NO_CONFIG_CMDS = {
+    'new',
+    'migrate',
+    'config',
+}
+
+
+_logger = logging.getLogger(__name__)
 _click_wrap_text = click.formatting.wrap_text
 
 
@@ -47,21 +58,47 @@ def _wrap_text(text: str, *a: Any, **kw: Any) -> str:
 
 click.formatting.wrap_text = _wrap_text
 
-ROOT_CONFIG = 'dipdup.yaml'
-CONFIG_RE = r'dipdup.*\.ya?ml'
 
-# NOTE: Do not try to load config for these commands as they don't need it
-NO_CONFIG_CMDS = {
-    'new',
-    'install',
-    'uninstall',
-    'update',
-    'migrate',
-    'config',
-}
+def _get_paths(
+    params: dict[str, Any],
+) -> tuple[list[Path], list[Path]]:
+    from dipdup.exceptions import ConfigurationError
+
+    config_args: list[str] = params.pop('config', [])
+    env_file_args: list[str] = params.pop('env_file', [])
+    config_alias_args: list[str] = params.pop('c', [])
+
+    config_paths: list[Path] = []
+    env_file_paths: list[Path] = []
+
+    if config_alias_args:
+        if config_args:
+            raise ConfigurationError('Cannot use both `-c` and `-C` options at the same time')
+        config_args = ['.', *(f'configs/dipdup.{name}.yaml' for name in config_args)]
+
+    config_args = config_args or ['.']
+
+    for arg in config_args:
+        path = Path(arg if arg != '.' else ROOT_CONFIG)
+        if not path.is_file():
+            raise ConfigurationError(f'Config file not found: {path}')
+        config_paths.append(path)
+
+    for arg in env_file_args:
+        path = Path(arg)
+        if not path.is_file():
+            raise ConfigurationError(f'Env file not found: {path}')
+        env_file_paths.append(path)
+
+    return config_paths, env_file_paths
 
 
-_logger = logging.getLogger(__name__)
+def _load_env_files(env_file_paths: list[Path]) -> None:
+    for path in env_file_paths:
+        from dotenv import load_dotenv
+
+        _logger.info('Applying env_file `%s`', path)
+        load_dotenv(path, override=True)
 
 
 def echo(message: str, err: bool = False, **styles: Any) -> None:
@@ -171,12 +208,11 @@ def _skip_cli_group() -> bool:
         ['schema'],
     )
     # NOTE: Simple helpers that don't use any of our cli boilerplate
-    is_script = args[0] in (
+    is_script_group = args[0] in (
         'report',
         'self',
     )
-    if not (is_help or is_empty_group or is_script):
-        _logger.debug('Skipping cli group')
+    if not (is_help or is_empty_group or is_script_group):
         return False
     return True
 
@@ -211,7 +247,7 @@ def _skip_cli_group() -> bool:
     '-C',
     type=str,
     multiple=True,
-    help='A shorthand for `-c . -c configs/dipdup.<name>.yaml',
+    help='A shorthand for `-c . -c configs/dipdup.<name>.yaml`',
     default=[],
     metavar='NAME',
 )
@@ -230,29 +266,9 @@ async def cli(ctx: click.Context, config: list[str], env_file: list[str], c: lis
     except AttributeError:
         _logger.warning("You're running an outdated Python 3.12 release; consider upgrading")
 
-    from dipdup.exceptions import ConfigurationError
     from dipdup.sys import set_up_logging
 
     set_up_logging()
-
-    if c:
-        if config:
-            raise ConfigurationError('Cannot use both `-c` and `-C` options at the same time')
-        config = ['.', *(f'configs/dipdup.{name}.yaml' for name in c)]
-    if not config:
-        config = [ROOT_CONFIG]
-
-    config_paths = [Path(file) for file in config]
-    env_file_paths = [Path(file) for file in env_file]
-
-    # NOTE: Apply env files before loading the config
-    for env_path in env_file_paths:
-        from dotenv import load_dotenv
-
-        if not env_path.is_file():
-            raise ConfigurationError(f'env file `{env_path}` does not exist')
-        _logger.info('Applying env_file `%s`', env_path)
-        load_dotenv(env_path, override=True)
 
     # NOTE: These commands need no other preparations
     if ctx.invoked_subcommand in NO_CONFIG_CMDS:
@@ -262,6 +278,11 @@ async def cli(ctx: click.Context, config: list[str], env_file: list[str], c: lis
     from dipdup.config import DipDupConfig
     from dipdup.exceptions import InitializationRequiredError
     from dipdup.package import DipDupPackage
+
+    # NOTE: Early config loading; some commands do it later
+    config_paths, env_file_paths = _get_paths(ctx.params)
+    # NOTE: Apply env files before loading the config
+    _load_env_files(env_file_paths)
 
     _config = DipDupConfig.load(
         paths=config_paths,
@@ -360,9 +381,10 @@ async def migrate(ctx: click.Context, dry_run: bool) -> None:
     from dipdup.config import DipDupConfig
     from dipdup.migrations.three_zero import ThreeZeroProjectMigration
 
-    # NOTE: Extract paths from arguments since we can't load config with old spec version
+    # NOTE: Late loading: can't load config with old spec version
     assert ctx.parent
-    config_paths: list[Path] = [Path(file) for file in ctx.parent.params['config']]
+    config_paths, env_file_paths = _get_paths(ctx.parent.params)
+    _load_env_files(env_file_paths)
 
     migration = ThreeZeroProjectMigration(tuple(config_paths), dry_run)
     migration.migrate()
@@ -414,7 +436,10 @@ async def config_export(
     from dipdup.config import DipDupConfig
     from dipdup.yaml import DipDupYAMLConfig
 
-    config_paths = [Path(c) for c in ctx.parent.parent.params['config']]  # type: ignore[union-attr]
+    # NOTE: Late loading; cli() was skipped.
+    config_paths, env_file_paths = _get_paths(ctx.parent.parent.params)  # type: ignore[union-attr]
+    _load_env_files(env_file_paths)
+
     if raw:
         raw_config, _ = DipDupYAMLConfig.load(
             paths=config_paths,
@@ -456,7 +481,9 @@ async def config_env(
     """
     from dipdup.yaml import DipDupYAMLConfig
 
-    config_paths = [Path(c) for c in ctx.parent.parent.params['config']]  # type: ignore[union-attr]
+    # NOTE: Late loading; cli() was skipped.
+    config_paths, env_file_paths = _get_paths(ctx.parent.parent.params)  # type: ignore[union-attr]
+    _load_env_files(env_file_paths)
 
     _, environment = DipDupYAMLConfig.load(
         paths=config_paths,